requirements.txt
google-cloud-speech==2.21.0
google-cloud-translate==3.11.1
google-generativeai==0.8.4
sounddevice==0.4.6
numpy==1.24.3
python-dotenv==1.0.0
scipy==1.11.4
flask==2.3.3
flask-cors==4.0.0
pydub==0.25.1
pymongo==4.6.0
PyJWT==2.8.0
werkzeug==2.3.7
librosa==0.10.1
soundfile==0.12.1
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
datasets==2.14.0
huggingface_hub>=0.19.0
accelerate>=0.24.0
requests==2.31.0
